requirements.txt
setuptools>=65.5.1
Django==4.2.7
django-environ==0.11.2
whitenoise==6.5.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
djangorestframework==3.16.0
django-cors-headers==4.7.0